refactor(#3880): 迁移模型枚举到 ginkgo/enums.py，统一包级导入
将8个枚举/常量类从模型文件迁移到 ginkgo/enums.py 统一管理，
同时将43个文件中的直接模型导入(from ginkgo.data.models.model_xxx)
修正为包级导入(from ginkgo.data.models)，从根本上解决 SQLAlchemy
mapper 初始化时 relationship() 字符串引用无法解析的问题。

枚举迁移: ExchangeType, EnvironmentType, AccountStatusType,
PermissionType, SubscriptionDataType, BrokerStateType,
DEPLOYMENT_STATUS, VALIDATION_STATUS

模型文件保留 re-export 以维持向后兼容。
