+++ src/ginkgo/data/models/model_live_account.py
@@ -10,81 +10,7 @@ from sqlalchemy import String, DateTime, Boolean, Text, Enum as SQLEnum
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 from ginkgo.data.models.model_mysqlbase import MMysqlBase
-
-
-class ExchangeType(str):
-    """交易所类型枚举"""
-    OKX = "okx"
-    BINANCE = "binance"
-
-    @classmethod
-    def from_str(cls, value: str) -> str:
-        """从字符串转换为枚举值"""
-        value = value.lower()
-        if value == "okx":
-            return cls.OKX
-        elif value == "binance":
-            return cls.BINANCE
-        else:
-            raise ValueError(f"Unknown exchange type: {value}")
-
-    @classmethod
-    def validate(cls, value: str) -> bool:
-        """验证交易所类型是否有效"""
-        return value in [cls.OKX, cls.BINANCE]
-
-
-class EnvironmentType(str):
-    """环境类型枚举"""
-    PRODUCTION = "production"
-    TESTNET = "testnet"
-
-    @classmethod
-    def from_str(cls, value: str) -> str:
-        """从字符串转换为枚举值"""
-        value = value.lower()
-        if value == "production":
-            return cls.PRODUCTION
-        elif value == "testnet":
-            return cls.TESTNET
-        else:
-            raise ValueError(f"Unknown environment type: {value}")
-
-    @classmethod
-    def validate(cls, value: str) -> bool:
-        """验证环境类型是否有效"""
-        return value in [cls.PRODUCTION, cls.TESTNET]
-
-
-class AccountStatusType(str):
-    """账号状态类型枚举"""
-    ENABLED = "enabled"
-    DISABLED = "disabled"
-    CONNECTING = "connecting"
-    DISCONNECTED = "disconnected"
-    ERROR = "error"
-
-    @classmethod
-    def from_str(cls, value: str) -> str:
-        """从字符串转换为枚举值"""
-        value = value.lower()
-        if value == "enabled":
-            return cls.ENABLED
-        elif value == "disabled":
-            return cls.DISABLED
-        elif value == "connecting":
-            return cls.CONNECTING
-        elif value == "disconnected":
-            return cls.DISCONNECTED
-        elif value == "error":
-            return cls.ERROR
-        else:
-            raise ValueError(f"Unknown account status type: {value}")
-
-    @classmethod
-    def validate(cls, value: str) -> bool:
-        """验证账号状态是否有效"""
-        return value in [cls.ENABLED, cls.DISABLED, cls.CONNECTING, cls.DISCONNECTED, cls.ERROR]
+from ginkgo.enums import ExchangeType, EnvironmentType, AccountStatusType  # noqa: F401 — re-export (#3880)
 
 
 class Base(DeclarativeBase):
