+++ src/ginkgo/trading/brokers/broker_manager.py
@@ -8,9 +8,10 @@ from datetime import datetime
 
 from ginkgo.libs import GLOG
 from ginkgo.data.containers import container
-from ginkgo.data.models.model_broker_instance import MBrokerInstance, BrokerStateType
-from ginkgo.data.models.model_portfolio import MPortfolio
-from ginkgo.data.models.model_live_account import AccountStatusType, ExchangeType
+from ginkgo.data.models import MBrokerInstance
+from ginkgo.enums import BrokerStateType
+from ginkgo.data.models import MPortfolio
+from ginkgo.enums import AccountStatusType, ExchangeType
 
 
 class BrokerManager:
